// src/app/api/swagger/models/api-application-dto-2.ts
/* tslint:disable */
/* eslint-disable */
import { ApiAdditionalServiceDto } from '../models/api-additional-service-dto';
import { ApiApplicationDocumentDto } from '../models/api-application-document-dto';
import { ApiApplicationItemDto } from '../models/api-application-item-dto';
import { ApiApplicationProductDto } from '../models/api-application-product-dto';
import { ApiApplicationSourceEnum } from '../models/api-application-source-enum';
import { ApiApplicationStateEnum } from '../models/api-application-state-enum';
import { ApiAuctionBidDto } from '../models/api-auction-bid-dto';
import { ApiAvailableVatTypeEnum } from '../models/api-available-vat-type-enum';
import { ApiContactInfoDto } from '../models/api-contact-info-dto';
import { ApiOrganizationRegistryDocumentDto } from '../models/api-organization-registry-document-dto';
import { ApiPaymentMethodEnum } from '../models/api-payment-method-enum';
import { ApiTradeBusinessFlowEnum } from '../models/api-trade-business-flow-enum';
import { ApiTradeStateEnum } from '../models/api-trade-state-enum';
import { ApiVatRateStateEnum } from '../models/api-vat-rate-state-enum';
export interface ApiApplicationDto2 {
  AccountNumber?: string | null;
  AdditionalServices?: Array<ApiAdditionalServiceDto> | null;
  AgreeWithCustomerConditions?: boolean;
  ApplicationDocuments?: Array<ApiApplicationDocumentDto> | null;
  ApplicationItems?: Array<ApiApplicationItemDto> | null;
  ApplicationSource?: ApiApplicationSourceEnum | null;
  ApplicationState?: ApiApplicationStateEnum;
  ApplicationStateDescription?: string | null;
  ApplicationStateId?: number | null;
  AuctionBids?: Array<ApiAuctionBidDto> | null;
  BankingDetailId?: number | null;
  CanDelete?: boolean;
  CanEdit?: boolean;
  CanWithdraw?: boolean;
  ConditionsOfPayment?: string | null;
  ContactInfo?: ApiContactInfoDto | null;
  CreatedByCustomer?: boolean;
  DefineCustomPriceForEachProduct?: boolean;
  DeliveryPlace?: string | null;
  DeliveryTerms?: string | null;
  EmailForReceipt?: string | null;
  EnableKtruSpecifications?: boolean;
  EnablePositionCharacteristics?: boolean;
  Id?: number;
  IncomplianceWithTheRequirementsReason?: string | null;
  /** @deprecated */IsLotPriceWithVat?: boolean;
  IsOrderApplication?: boolean;
  IsPriceEnteredWithVat?: boolean;
  IsPublicOffer?: boolean;
  IsQuotationWithVat?: boolean | null;
  IsRequirementsConfirmationCheck?: boolean;
  IsSelfDelivery?: boolean;
  IsSellerDelivery?: boolean;
  IsTransportCompanyDelivery?: boolean;
  NeverExpired?: boolean;
  OfferExpiryDate?: string | null;
  OrganizationRegistryDocuments?: Array<ApiOrganizationRegistryDocumentDto> | null;
  PayCommissionReportGuid?: string;
  PaymentMethod?: ApiPaymentMethodEnum;
  Price?: number | null;
  PriceWithVat?: number | null;
  Products?: Array<ApiApplicationProductDto> | null;
  PublicationDate?: string | null;
  RateVat?: ApiAvailableVatTypeEnum | null;
  ReductionCoefficient?: number | null;
  ResultPosition?: number | null;
  RowVersion?: number;
  SelfDeliveryPlace?: string | null;
  SellerDeliveryPrice?: number | null;
  ShipmentPoint?: string | null;
  SubmitterId?: string | null;
  SubmitterName?: string | null;
  SumVat?: number | null;
  TenantSource?: number | null;
  TradeBusinessFlow?: ApiTradeBusinessFlowEnum;
  TradeId?: number;
  TradeLotId?: number;
  TradeOrganizerId?: number | null;
  TradeState?: ApiTradeStateEnum;
  TransportCompanyDeliveryPrice?: number | null;
  /** @deprecated */VatRate?: number | null;
  /** @deprecated */VatRateDescription?: string | null;
  /** @deprecated */VatRateState?: ApiVatRateStateEnum;
  WasRejected?: boolean;
}